Обработка корзины и заказов вынесены из ProductService в CartService и OrderService
