+++ src/main/java/io/github/tasoula/intershop/controller/CartController.java
@@ -2,7 +2,7 @@ package io.github.tasoula.intershop.controller;
 
 import io.github.tasoula.intershop.dto.ProductDto;
 import io.github.tasoula.intershop.interceptor.UserInterceptor;
-import io.github.tasoula.intershop.service.ProductService;
+import io.github.tasoula.intershop.service.CartService;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
@@ -11,7 +11,6 @@ import org.springframework.web.bind.annotation.*;
 
 import java.math.BigDecimal;
 import java.util.List;
-import java.util.Optional;
 import java.util.UUID;
 
 @Controller
@@ -19,9 +18,9 @@ import java.util.UUID;
 public class CartController {
 
     //todo как-то разделить продукты и cart_item здесь и в сервисах
-    private final ProductService service;
+    private final CartService service;
 
-    public CartController(ProductService service) {
+    public CartController(CartService service) {
         this.service = service;
     }
 
@@ -36,24 +35,6 @@ public class CartController {
         return "cart.html";
     }
 
-    @PostMapping("items/{id}")
-    public ResponseEntity<Integer> changeCartQuantity(HttpServletRequest request,
-                                                      @PathVariable("id") UUID id,
-                                                      @RequestParam("action") String action,
-                                                      Model model) {
-        UUID userId = UUID.fromString((String) request.getAttribute(UserInterceptor.USER_ID_COOKIE_NAME));
-
-        int newQuantity = 0;
-        if(action.equals("delete")){
-            service.deleteCartItem(userId, id);
-        }
-        else {
-            newQuantity = service.changeCartQuantity(userId, id, action.equals("plus") ? 1 : -1); // Get the new quantity from the service
-        }
-
-        return ResponseEntity.ok(newQuantity); // Return the updated quantity
-    }
-
     @GetMapping("total")
     private ResponseEntity<BigDecimal> getTotal(HttpServletRequest request){
         UUID userId = UUID.fromString((String) request.getAttribute(UserInterceptor.USER_ID_COOKIE_NAME));
@@ -66,10 +47,21 @@ public class CartController {
         return ResponseEntity.ok(service.isEmpty(userId));
     }
 
-    @PostMapping("buy")
-    public String buy(HttpServletRequest request, Model model) {
+    @PostMapping("items/{id}")
+    public ResponseEntity<Integer> changeProductQuantityInCart(HttpServletRequest request,
+                                                               @PathVariable("id") UUID productId,
+                                                               @RequestParam("action") String action) {
         UUID userId = UUID.fromString((String) request.getAttribute(UserInterceptor.USER_ID_COOKIE_NAME));
-        Optional<UUID> orderId = service.createOrder(userId);
-        return orderId.map(uuid -> "redirect:/orders/" + uuid + "?newOrder=true").orElse("redirect:/cart/items");
+
+        int newQuantity = 0;
+        if(action.equals("delete")){
+            service.deleteCartItem(userId, productId);
+        }
+        else {
+            newQuantity = service.changeProductQuantityInCart(userId, productId, action.equals("plus") ? 1 : -1);
+        }
+
+        return ResponseEntity.ok(newQuantity);
     }
+
 }
