+++ src/main/java/io/github/tasoula/intershop/model/Product.java
@@ -38,4 +38,8 @@ public class Product {
 
     @Column(name = "stock_quantity", nullable = false)
     private int stockQuantity;
+
+    public Product(UUID productId) {
+        this.id = productId;
+    }
 }
